reorganize locales file structure
