+++ locales/config/languages.config.json
